add ut local torch binary install

+++ scripts/create_conda_env.sh
@@ -95,8 +95,19 @@ if [[ ${run_ut} != '' ]]; then
     fi
 
     # Install PyTorch
-    pip install torch==${pytorch_version} -f https://download.pytorch.org/whl/torch_stable.html
-    pip install torchvision==${torchvision_version} -f https://download.pytorch.org/whl/torch_stable.html
+    torch_whl_path=/tf_dataset/pytorch/pypi
+    torch_whl=${torch_whl_path}/${python_version}/torch-${pytorch_version}-*.whl
+    if [ -f ${torch_whl} ]; then
+        pip install ${torch_whl}
+    else
+        pip install torch==${pytorch_version} -f https://download.pytorch.org/whl/torch_stable.html
+    fi
+    torchvision_whl=${torch_whl_path}/${python_version}/torchvision-${torchvision_version}-*.whl
+    if [ -f ${torchvision_whl} ]; then
+        pip install ${torchvision_whl}
+    else
+        pip install torchvision==${torchvision_version} -f https://download.pytorch.org/whl/torch_stable.html
+    fi
 
     # Install MXNet
     if [ ${mxnet_version} == '1.6.0' ]; then
